requirements.txt
flask
gunicorn
python-dotenv
aiogram
aiohttp
aiofiles
requests
loguru
